+++ packages/quick_actions/example/android/app/src/main/AndroidManifest.xml
@@ -12,6 +12,14 @@
               android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection"
               android:hardwareAccelerated="true"
               android:windowSoftInputMode="adjustResize">
+    </activity>
+    <activity android:name="dev.flutter.plugins.quickactionexample.MainActivity"
+        android:launchMode="singleTop"
+        android:theme="@style/LaunchTheme"
+        android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection"
+        android:hardwareAccelerated="true"
+        android:windowSoftInputMode="adjustResize">
+    </activity>
       <meta-data
           android:name="io.flutter.app.android.SplashScreenUntilFirstFrame"
           android:value="true"/>
